fix: auto mount after formatted.
In some file systems, path file attribute lookups will be delayed in
returning if you immediately enter a device that has just been
formatted. And because of this, some actions in context menu is disabled
due to the real read write permission is not obtained when context menu
triggered.
Auto mount after formatted, so that there could be a large interval user
first enter into device, and the permission attributes could be updated
correctly.

Log: as above.

Bug: https://pms.uniontech.com/bug-view-296699.html